fixed bugs and design problems and refactor
