+++ .eslintrc.json
@@ -133,16 +133,6 @@
     {
       "files": ["examples/**/*"],
       "rules": {
-        "@typescript-eslint/no-use-before-define": [
-          "error",
-          {
-            "functions": true,
-            "classes": true,
-            "variables": true,
-            "enums": true,
-            "typedefs": true
-          }
-        ],
         "import/no-anonymous-default-export": [
           "error",
           {
@@ -364,7 +354,6 @@
     "react/require-render-return": "error",
     "react/style-prop-object": "warn",
     "react-hooks/rules-of-hooks": "error",
-    // "@typescript-eslint/non-nullable-type-assertion-style": "warn",
     "@typescript-eslint/prefer-as-const": "warn",
     "@typescript-eslint/no-redeclare": [
       "warn",
