+++ app/[lang]/arbswap/(protected)/dashboard/layout.tsx
@@ -1,23 +1,17 @@
 import React from 'react'
 
-import AppBar from '@/components/CustomNavbar'
+import ArbSwapNavbar from '@/components/CustomNavbar/ArbSwap'
 
 type RootLayoutProps = Readonly<{
   children: React.ReactNode
-  form: React.ReactNode
   params: { lang: string }
 }>
 
-export default function DashboardLayout({
-  children,
-  form,
-  params
-}: RootLayoutProps) {
+export default function DashboardLayout({ children, params }: RootLayoutProps) {
   return (
     <section className="min-h-screen overflow-hidden bg-gradient-to-t from-purple-950 via-slate-900 to-slate-950">
-      <AppBar locale={params.lang} />
+      <ArbSwapNavbar locale={params.lang} />
       {children}
-      {form}
     </section>
   )
 }
