+++ components/App/ArbSwap/Dashboard/Offers/OfferHistory/index.tsx
@@ -1,5 +1,6 @@
+import { Offer } from '@/types/dashboard'
+
 import React, { useEffect, useState } from 'react'
-import { Offer } from '@/constants/Offer'
 
 import Received from '@/components/App/ArbSwap/Dashboard/Offers/OfferHistory/Received'
 import Sent from '@/components/App/ArbSwap/Dashboard/Offers/OfferHistory/Sent'
