+++ components/App/ArbSwap/Dashboard/PaymentMethod.tsx
@@ -5,13 +5,13 @@ import React from 'react'
 import Button from '@/components/Common/Button'
 import Modal from '@/components/Common/Modal'
 import getFormattedDateTime from '@/hooks/useCurrentDate'
-import { useModalStore } from '@/hooks/useStore'
+import { modalStore } from '@/hooks/useStore'
 
 function PaymentMethod() {
   const modalKey = 'paymentMethod'
   const otherModalKey = 'pricing'
 
-  const { toggleModal } = useModalStore()
+  const { toggleModal } = modalStore()
 
   const handleNext = () => {
     toggleModal(modalKey)
