+++ components/CustomFooter/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 import { FooterSocials } from '@/constants'
-import { footerDictionary } from '@/localesContent'
+import { footerDictionary } from '@/constants/localesContent'
 
 interface CustomFooterProps {
   readonly locale: string
