--- constants/Listing.ts
@@ -1,10 +1,0 @@
-export interface Listing {
-  id: string
-  user_id: string
-  seller_username: string
-  amount: number
-  total_price: number
-  currency: string
-  is_deleted: boolean
-  is_active: boolean
-}
