+++ types/dashboard.ts
@@ -1,3 +1,14 @@
+export interface Listing {
+  id: string
+  user_id: string
+  seller_username: string
+  amount: number
+  total_price: number
+  currency: string
+  is_deleted: boolean
+  is_active: boolean
+}
+
 export interface Offer {
   user_id: string
   status: string
