+++ types/locales.ts
@@ -1,0 +1,170 @@
+export interface ArbswapNavbar {
+  aboutUs: string
+  createOffer: string
+  wallet: string
+  support: string
+  login: string
+  walletConnected: string
+  goToDashboard: string
+}
+
+export interface NexBridgeNavbar {
+  home: string
+  connectWallet: string
+  pricing: string
+  dashboard: string
+}
+
+export interface Navbar {
+  aboutUs: string
+  ourTeam: string
+  contact: string
+  joinUs: string
+}
+
+export interface About {
+  aboutNexarb: string
+  description: string
+  manyUsefulComponents: string
+  manyUsefulComponentsDesc: string
+  responsiveLayout: string
+  responsiveLayoutDesc: string
+  retinaReady: string
+  retinaReadyDesc: string
+}
+
+export interface ArbSwapHome {
+  introHeader: string
+  introDescription: string
+  getStarted: string
+  buySOL: string
+  sellSOL: string
+  amount: string
+  searchForOffers: string
+  exchangeRates: string
+  userWorldwide: string
+  countrySupported: string
+  cryptoTransactions: string
+  assetsOnPlatform: string
+  sendAndRecieve: string
+  sendAndRecieveDesc: string
+  stepsHeader: string
+  quickSignUp: string
+  quickSignUpDesc: string
+  securitySetup: string
+  connectWallet: string
+  startTrading: string
+  startTradingDesc: string
+  solanaWorld: string
+  solanaWorldDescription: string
+  signUpNow: string
+  confidentiality: string
+  confidentialityDesc: string
+}
+
+export interface NexBridgeHome {
+  description: string
+  crossChain: string
+}
+
+export interface Home {
+  digitalSolutions: string
+  digitalSolutionsDesc: string
+  letsMeet: string
+  ourWorks: string
+  ourWorksDesc: string
+  startupCrew: string
+}
+
+export interface NexBridgeDashboard {
+  compose: string
+  bridgeOptions: string
+  edit: string
+  delete: string
+  from: string
+  to: string
+  personalData: string
+  startBridge: string
+  information: string
+}
+
+export interface Login {
+  emailOrPhoneNumber: string
+  enterUsername: string
+  password: string
+  enterPassword: string
+  rememberThisDevice: string
+  login: string
+  loginWithGoogle: string
+  dontYouHaveAnAccount: string
+  signUpHere: string
+  forgotPassword: string
+  verifyHere: string
+  connectWallet: string
+}
+
+export interface Register {
+  enterEmail: string
+  choosePassword: string
+  enterPassword: string
+  passwordAgain: string
+  confirmPassword: string
+  acceptTermsConditions: string
+  next: string
+}
+
+export interface RegisterWallet {
+  connectPhantomWallet: string
+  connectWallet: string
+  walletConnected: string
+  register: string
+}
+
+export interface ForgotPassword {
+  enterEmail: string
+  sendCode: string
+  verificationCode: string
+  enterVerificationCode: string
+  dontYouGetCode: string
+  sendAgain: string
+  verify: string
+  newPassword: string
+  enterPassword: string
+  newPasswordAgain: string
+  login: string
+}
+
+export interface ArbSwapFooter {
+  dummyText: string
+  products: string
+  overview: string
+  solutions: string
+  pricing: string
+  customers: string
+  company: string
+  about: string
+  inverstorRelations: string
+  jobs: string
+  press: string
+  blog: string
+  support: string
+  contact: string
+  documentation: string
+  chat: string
+  faq: string
+  legal: string
+  termsOfService: string
+  privacyPolicy: string
+  cookieSettings: string
+  copyright: string
+}
+
+export interface Footer {
+  keepInTouch: string
+  description: string
+  yourName: string
+  inputField: string
+  yourMessage: string
+  send: string
+  copyright: string
+}
